Revert "media: Change VideoFrame's stride type to size_t"
This reverts commit abc3810f20082e06b47fc70fa1da35786a63dc14.

Reason for revert: Build failed here: https://ci.chromium.org/ui/p/chromium/builders/ci/chromeos-arm-generic-rel/135339/overview

Original change's description:
> media: Change VideoFrame's stride type to size_t
>
> We don't support negative strides anymore, so no reason to keep it
> in a signed type.
>
> Bug: 338570700
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020592
> Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1383763}

Bug: 338570700
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6026497
Commit-Queue: Lei Zhang <thestig@chromium.org>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1383777}

+++ media/base/color_plane_layout.h
@@ -15,20 +15,21 @@
 namespace media {
 
 // Encapsulates a color plane's memory layout: (stride, offset, size)
-// stride: the number of bytes, including padding, used by each row of a plane.
+// stride: in bytes of a plane. Note that stride can be negative if the image
+//         layout is bottom-up.
 // offset: in bytes of a plane, which stands for the offset of a start point of
 //         a color plane from a buffer FD.
 // size:   in bytes of a plane. This |size| bytes data must contain all the data
 //         a decoder will access (e.g. visible area and padding).
 struct MEDIA_EXPORT ColorPlaneLayout {
   ColorPlaneLayout();
-  ColorPlaneLayout(size_t stride, size_t offset, size_t size);
+  ColorPlaneLayout(int32_t stride, size_t offset, size_t size);
   ~ColorPlaneLayout();
 
   bool operator==(const ColorPlaneLayout& rhs) const;
   bool operator!=(const ColorPlaneLayout& rhs) const;
 
-  size_t stride = 0;
+  int32_t stride = 0;
   size_t offset = 0;
   size_t size = 0;
 };
